+++ embassy-stm32/src/rtc/mod.rs
@@ -1,4 +1,6 @@
 //! Real Time Clock (RTC)
+#[cfg(rtc_v3)]
+mod alarm;
 mod datetime;
 
 #[cfg(feature = "low-power")]
@@ -12,6 +14,8 @@ use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 #[cfg(feature = "low-power")]
 use embassy_sync::blocking_mutex::Mutex;
 
+#[cfg(rtc_v3)]
+pub use self::alarm::{Alarm, AlarmDate, RtcAlarmMatch};
 use self::datetime::{day_of_week_from_u8, day_of_week_to_u8};
 pub use self::datetime::{DateTime, DayOfWeek, Error as DateTimeError};
 use crate::pac::rtc::regs::{Dr, Tr};
@@ -149,6 +153,15 @@ impl Rtc {
     pub fn new(_rtc: Peri<'static, RTC>, rtc_config: RtcConfig) -> Self {
         #[cfg(not(any(stm32l0, stm32f3, stm32l1, stm32f0, stm32f2)))]
         crate::rcc::enable_and_reset::<RTC>();
+
+        #[cfg(rtc_v3)]
+        {
+            use crate::interrupt::typelevel::Interrupt;
+            <RTC as crate::rtc::SealedInstance>::AlarmInterrupt::unpend();
+            unsafe {
+                <RTC as crate::rtc::SealedInstance>::AlarmInterrupt::enable();
+            }
+        }
 
         let mut this = Self {
             #[cfg(feature = "low-power")]
@@ -297,6 +310,12 @@ trait SealedInstance {
     #[cfg(feature = "low-power")]
     type WakeupInterrupt: crate::interrupt::typelevel::Interrupt;
 
+    #[cfg(rtc_v3)]
+    const EXTI_ALARM_LINE: Option<usize>;
+
+    #[cfg(rtc_v3)]
+    type AlarmInterrupt: crate::interrupt::typelevel::Interrupt;
+
     fn regs() -> crate::pac::rtc::Rtc {
         crate::pac::RTC
     }
@@ -312,6 +331,10 @@ trait SealedInstance {
     /// The registers retain their values during wakes from standby mode or system resets. They also
     /// retain their value when Vdd is switched off as long as V_BAT is powered.
     fn write_backup_register(rtc: crate::pac::rtc::Rtc, register: usize, value: u32);
+
+    fn write<F, R>(init_mode: bool, f: F) -> R
+    where
+        F: FnOnce(crate::pac::rtc::Rtc) -> R;
 
     // fn apply_config(&mut self, rtc_config: RtcConfig);
 }